feat: implement validate qr endpoint with signature
